event-auditor: split ebpf (Go) logic between files
